修改服务配置弹窗代码结构
